Add smooth enabling/disabling for buttons #70
One minor TODO here: for basic inputs (like text fields and the
colorization radio buttons), adding this transition doesn't do
anything on enabling/disabling. This seems to be since these tags
have their style changes built in to the disabled attribute, I guess?

Not sure if there's a way to change that. But it's the sort of thing
that seems super super minor. I'll keep #70 open, but probably best
to focus on other stuff for the time being.