Differentiate finishing candidate border color
...from the added-to-path-during-finishing node/node group color.
This is because it's possible to loop back to a previously "finished"
node during the finishing process due to cycles in the graph -- so
we want to make sure that the user can see that a given node or node
group is again eligible to be added to the path.

I thought I'd taken care of this already? Guess not, though. It looks
like I accidentally reverted a solution to this problem in commit
030847227d2afedf9e1655237e136dd8876dc0a4 due to not remembering that
it's a valid case for a node/node group to be added to a given
finishing path multiple times.

In any case, this should be fixed.