Explicitly lbl attributions in xdot2cy.js; add 2
Towards the practice of labelling all in-code attributions with the
"ATTR:" prefix, in order to make finding them easier.

Also added two references that weren't there but should be (the
Cytoscape.js thing was previously stated in the README, and I don't
think the File API one was stated anywhere but I remember reading
that article). Should've had these in a while ago.