Add directory support for #113
Supports the TSV filename\tdescription files as an optional
argument.

More testing needed before I'll close #113, but this is looking
good.

(Note that I didn't finish generating august1_before.db a while
back, so it's removed temporarily from the index demo. Making a note
to add it back later.)

Also note that it looks like the edge count I had manually put in the
demo for the GML E. coli graph was slightly wrong. Kind of neat that
the demo-populating tool is already fixing errors.