Include note re license copyright years in README
This should make the License, Demo, and Contact sections of the
README all match with their respecive sections in the wiki. I don't
expect these sections to need to be updated much in the future, so
I'm ok with duplicating them like this; they differ from the
acknowledgements section a bit in that way, in that I might update
the acknowledgements section periodically in the future as
MetagenomeScope uses more and more dependencies.