Add -spqr option to py script & partial func #49
Right now, the -spqr option:
-Controls if the SPQR layouts are done (this is the main thing I
 wanted this feature to control, since performing these layouts can
 take a lot of time)
-Sets a new "spqr_given" value in the assembly table (works similarly
 to the dna_given and repeats_given values)

Still some work to be done on this. There's some SPQR stuff that's
done in the preprocessing script that should just be not done if
-spqr isn't passed. (Will make the script a bit faster in that case.)
Also, the viewer interface should look at the spqr_given field for
a database and toggle either the disabledness or visibility of
the SPQR component drawing controls accordingly, if needed.