Make more SPQR computations conditional on -spqr
In particular, some of the logistics re: links/bicmps files, calling
the SPQR script, processing component_*.info and spqr*.gml files,
"single" component detection, scaling contigs in single components,
and printing the standard mode layout time separately from total
layout time.

It's a lot of stuff, but I think that about covers it! Next up is
making the viewer interface respond correctly to .db files without
SPQR data. (It's already doing a decent job, but it should
1) hide the SPQR component layout controls, and
2) hide the SPQR cc count and bicomponent count sections of the asm.
   info table, since those are just listed as 0 in the .db if -spqr
   is not passed.

Noticed a weird bug with the sample_LastGraph (when -spqr is passed)
where one of the conclude_msg()s is on a separate line after the
explicit view layout time is printed. I looked into it, and this
error was around before I started working on #49 earlier today --
so it's a separate bug I've gotta fix. (Now it's its own thing on
GitHub as #114.)