Make id="demoDir" the TAG in #113 script for -hd
This keeps the HTML easy to read and avoids reliance on an
extraneous boolean attribute.

This method in the populate_demo script relies on the id="demoDir"
and data-mgscdbdirectory attribute being on the same line of the
file. Now that I think about it, I should really add something to
index.html specifying that.

Also added a "continue" just to ensure that the DB_LIST_START_TAG
doesn't mess with this if it's somehow on the same line as the
HOST_DB_DIR_TAG (although if that somehow happens things are probably
severely messed up in the index.html file anyway).