Expose autofinishing max zoom level: close #116
There seems to be a weird small style bug, where the hidden div
with the max zoom level input still takes up a tiny amount of
vertical space. Not sure re: how to address that, and I figure it'd
be silly to delay this commit because of that, so I just noted it in
the code for future reference.