Add selection of searching w/ IDs vs lbls #93
This addresses the "labels not working" bug in #93, since it removes
the retrospectively-silly reliance on prefixes to determine if
something is a label or ID.

A TODO is that, due to _ele2parent, this is actually *too*
exhaustive, and will find parent clusters of nodes if their ID is
passed in label search mode (or vice versa) -- which isn't the
desired behavior, since one node's ID could ostensibly clobber
another node's label (or vice versa).

Worth noting that this commit effectively deprecates support for
searching for edges/clusters. Which is fine, since those were already
hard to search for anyway (and I don't think supporting searching for
them is super important). But it might be worth (another possible
TODO) going through the search function and removing the ability to
identify clusters or edges (perhaps by limiting the scope of the
query data to search over -- and this might actually speed up
searching).

Also want to adjust the placeholder text based on the search mode.
A third TODO, I guess.