Add fn for getting num. IDs from FASTG node decls
Right now it's used for parsing node references in user-bubble and
user-pattern files, but it'd be a good idea to apply this to the
actual parsing code as well. Making this function work with other
sorts of IDs would be a good idea, I think.

I suppose that #66 being implemented would sort of make this function
obselete, however. This is a decent stopgap, though.