nvm don't bother with package stuff for now
Over an hour of looking into this shows that getting relative
imports to work properly both when a file is run as a module and
as a script seems to require some complicated, bizarre solutions.
It's the sort of thing I guess I'll have to get around to fixing
eventually when we port MetagenomeScope to Python 3, but for
the time maintaining the prior import system (implicit relative
imports) is fine, if not ideal.