Add (tentative) Webstrates version of index.html
In that it doesn't reference any local files -- all of the
JS/CSS/resources should be loaded via URLs, either from CDNs or from
the MetagenomeScope demo site.

This lets this HTML be plopped straight into a "webstrate" as a
stand-alone thing. You could also really put it anywhere, since it
should be an entirely stand-alone application.

A note: I'd like to eventually automate the process of creating this
file; see issue #129 for details on that.

The next TODO, I think, is getting this to apply canvas events across
instances. That'd facilitate directly collaborative visualization in
MetagenomeScope, which would be..... really cool.