Make -sp invocation of save_aux_file use warnings
Instead of having warnings=False. Since whether or not -sp generates
any output has no effect on the resulting .db file's generation, we
leave warnings=True when we call save_aux_file() -- so if one of the
files -sp generates can't be written, we just give the user a warning
explaining that and move on (instead of stopping execution of the
preprocessing script).

I *guess* this behavior could be configurable by the user if there's a
demand for that -- e.g. if layout takes forever, and users don't want to
have to repeat a run of the preproceessing script to generate the -sp
files -- but I don't see that being a very likely use case. (Also, the
output of -sp can probably just be extracted from the .db file after the
fact anyway. Shouldn't be too hard to write a Python script that does
that, if some user ends up in that scenario.)