Specify python2.7 in collate.py shebang
Makes the system disambiguate between Python 2 and Python 3.

Also, as detailed here: https://stackoverflow.com/q/45046409, an
advantage of specifying python2.7 in particular is that if no
python2.7 is set up, then the user will get that specific error
(instead of a Python 3 error about Python 2-specific syntax in the
preprocessing script).