Remove unneeded IE compatibility <meta> tag
See https://github.com/twbs/bootstrap/issues/21476 and
https://stackoverflow.com/q/6771258 for a thorough overview.

From the looks of it, this tag isn't needed for IE 10+, and IE 10
is the oldest version of IE that the dependencies of the viewer
interface support as of writing (per
https://github.com/farbelous/bootstrap-colorpicker/issues/271).

So there's no need to include this tag.

Also for posterity here's a list of all of the IE version support
for the viewer interface's current dependencies (I was going to
include this in a comment in the viewer interface that justified
keeping the tag in, but then I realized that this tag wasn't even
needed for IE 10 so I scrapped the comment) --

d3 v4: IE9+
    (https://github.com/d3/d3/wiki#supported-environments)
Cytoscape.js: IE9+ [supports canvas and most of ECMAScript 5]
    (http://js.cytoscape.org/#introduction/factsheet)
jQuery: IE9+
    (https://jquery.com/browser-support/)
bootstrap 3: IE 8-11
    (https://getbootstrap.com/docs/3.3/getting-started/#support)
bootstrap colorpicker: IE >= 10
    (https://github.com/farbelous/bootstrap-colorpicker/issues/271)

Couldn't find much data on sql.js browser support aside from that
old browsers probably aren't a good idea with it, so using IE
in general with sql.js is probably not feasible. (See:
https://github.com/kripken/sql.js/issues/238,
https://github.com/kripken/sql.js/issues/239)