Only use overlap=false on full SPQR layout&not -nt
Since overlap doesn't apply to dot (per
http://www.graphviz.org/doc/info/attrs.html#a:overlap), there's no
need to include it in GRAPH_STYLE for our invocations of dot.

We might want to add overlap=false (if -nt not passed) to the other
invocations of sfdp (in SPQRMetaNode.layout_isolated() and
Bicomponent.implicit_layout_isolated()), but let's see how this goes.

This should resolve the problems in #135 once that one Graphviz
error is fixed