Fix backwards compatibility issue with old DBs
Old .db files don't have a too_large column in their components,
so they were all failing. A simple fix for now is checking for if
an error is thrown and, if so, just acting like too_large is false
(which is the default behavior anyway).

there shouldn't be any worries re: needing to free isTooLargeStmt,
since it won't even be defined (the error from sql.js seems to occur
before the variable is created).