Move viewer int. tests to within the VI HTML #76
Since using iframes to embed MetagenomeScope within a mocha instance
-- or vice versa -- doesn't seem like it'll be doable in any useful
way due to CORS.

Of course, this test functionality shouldn't be included in most
distributions of MetagenomeScope's viewer interface to save on space.
Eventually I'll write a script or something that just inserts this
testing dialog into a new HTML page and runs the tests in that. But
there isn't any harm (aside from the extra bandwidth) of keeping these
tests in the viewer interface, anyway -- the functionality is hidden
from the normal user, and being able to test the V.I. might mess up the
graph state but if you're messing with the browser console then you've
pretty solidly already made it clear that you're just messing with
MetagenomeScope at that point. It's just doing client-side code stuff,
so it shouldn't pose a security risk either.