Update minification script shebang + add comments
The comments are regarding the aliasing of gsed over sed for systems
with gsed available. I was concerned earlier about how this might
cause errors for users (e.g. they go to use "sed" and are confused
as to why it maps to "gsed" instead), but it doesn't look like fish
preserves aliases (which from fish's perspective are just functions)
after the user runs a fish script -- so that shouldn't pose an issue.

[ci skip] since this doesn't actually hit any of MetagenomeScope's
code directly. Although maybe incorporating this into the travis
script is a good idea? something to think about