Add draft (curr. unused) rotation stuff tests #76
In order to use these, we need to first load in a test DB file to
the viewer interface. Not sure of the best way to do this --
maybe we could save some small DB files in memory as Uint8Arrays,
or something like that.

Not sure if we could find ways to test the DB loading (either
loadHostedDB() or loadLocalDB()) -- will look into it.