ENH: Port fish scripts to bash; remove Electron
SHEESH why did I write these in fish in the first place.
The bash script porting closes #149!

I ended up removing the very basic in-progress electron/ version,
because it added *a lot* of complexity to the minification process
-- and I'm not confident that it's actually useful to keep around,
until we go ahead with #5 and make it actually harness the node.js
library. If/when I get around to doing that, I think it'll require
the sort of changes that go beyond a few "sed" commands.

Also, I updated the copyright years in the minified attributions
(#150) -- need to actually change these throughout. Also I guess
we should include 2016 in this range? TBH I don't think we'd have
a problem with just "2016--".