MAINT: Split up graph_objects.py into a subpackage
...containing the various classes.

Also introduced a new AssemblyGraph class, which I'll use for
parsing stuff.

The AssemblyGraph class will delegate parsing to NetworkX (for GMLs)
and gfapy (for GFAs), taking work off our plate. That means that the
only parsers we need to write for this are for LastGraph and FASTG.

(I updated setup.py to add nx and gfapy as dependencies -- currently
unused but will be soon.)