TST: Add tenative GFA2 test; acct for "loop" case
the loop case is where an edge's complement is equal to itself
(e.g. 2+ -> 2- from loop.gfa c/o Shaun Jackman). forgot to account
for that.

the gfa2 test will fail due to gfa2 edges not having a complement
method as of now. probs possible to dev a workaround, one way or
another -- will work on that tomorrow