TST/BUG: Add cyclic chain chain tests; fix bug!
The bug is in the ...ambiguous test. I think this was actually a problem
with MetagenomeScope's old codebase, wow. (I mean it only would've
resulted in some wonky cyclic chains being represented as chains
instead... not the largest problem, but still a sign of how adding tests
in makes it easier to sleep at night ;)