TST: Fix an "intervening graph" test
Problem was that the body of a for loop just straight-up wasn't being
visited in ..._harder() because I used the wrong set operator (should've
used | instead of &). To make this part clearer, I just rewrote the list
of starting nodes to try explicitly.

as of writing it looks like this brings up the chain code to 100%
coverage! gonna move on now.