TST/MNT: add basic bubble unit tests; add comments
need to think about how to represent bubbles of chains in this new
paradigm -- I don't think collapsing a chain is really useful? So
maybe acknowledging that chains won't be collapsed is ok, i guess?
or adjusting stuff so that chains are always collapsed before running
the bubble /frope detection for a given level