MNT: Finish rewriting simple bubble detection code
So the way I think it makes the most sense to do this is to make
this only detect simple bubbles where each middle path contains a
single node.

Now, I know what you're thinking -- "why would you remove
functionality like that?" After all, the old version of MgSc allowed
for "long" paths in the middle of bubbles (and frayed ropes, for that
matter).

The answer is that this isn't really removing functionality. When
we do iter. decomp., keeping bubble detection simple will force the
code to collapse chains (if they exist) inside bubbles first, and
then we'll collapse the corresponding bubbles later on.

The advantage of this is that this preserves the "chains" as distinct
patterns inside bubbles, and ensures that we're not inconsistent as
to whether not we tag these as chains or leave them as just parts of
bubbles. Doing things this way is a lot simpler to implement than
something fancy where the bubble code produces a list of chains *and*
bubble info, or something...