Unique int node ids; implement #84 dupe node soln
This is notably kinda funky. In particular, it's possible for
dupe edges to compound in weird ways, as shown in the test case.

i think it might also be possible for frayed ropes to infinitely
create new ones...?

I think the way around this is having dupe edges temporarily? keep
track of the pattern that caused them to be created, and disallowing
patterns being constructed where all edges are dupe edges
created from the same pattern.

This would address the most egregious cases, although this would
"break" some things in the sense that the current bubble-finding /
frayed rope finding algorithms assume that chains have already been
collapsed -- so if we have something like

    2
   / \
  1   4
 / \ / \
0   3   6
 \     /
  \-5-/

then it'll get turned into

  7=B=8
 /     \
0       6
 \     /
  \-5-/

... and then this won't get tagged as a bubble, even though it really
should be!!! We could modify the bubble/frayed rope code to check for
chains, though. that would honestly not be too bad, although it might
start to be weirdly inconsistent............. maybe we'd ONLY allow
chains in bubbles when all of the edges in the chain are duplicates?

Or here's a better idea. When we create a chain, if all of the edges
within it are duplicates arising from the same pattern, then we mark
it specifically as a "temporary" chain. This means that, when it would
come time to create a pattern containing this chain, we wouldn't include
the chain itself -- we'd just include the contents of the chain
directly, as if the chain had never existed.
OH, and if we get to the end of the decomposition and the temporary
chain is still around, then we can also just remove the temporary chain
and leave its contents uncollapsed.

We may even be able to remove the duplicate nodes in this case -- since
it'd be clear that they aren't being used for anything meaningful.

Yeah i thiiiink this might be the way forward? i'm uncertain if removing
the duplicate nodes in the tmp-chain-case will be always possible, but
at the very least marking this as a tmp chain and removing it is ok.
ACTUALLY if we mandate that duplicate nodes are ONLY created for
bubbles, then that'll mean that there can only be two duplicate edges
per bubble: so if both are included in a chain then we know for a FACT
that the duplicate edges are useless in this case.

Other poss solutions:

1) disallow starting duplicates to be used as start of another pattern,
   and disallow ending duplicates to be used as end of another pattern
   (would need to keep track of starting vs ending duplicate status, ez)

2) disallow patterns being created where the start and end nodes are
   both duplicates.

both of these two are bad, because there are legitimately useful
   cases where this could happen -- e.g.

    dupe == bubble == bubble == dupe