GET #84 BUBBLE CHAIN STUFF KINDA WORKING
:D:D:D

So, it isn't perfect -- there will be cases where this does weird
and suboptimal things, e.g. prematurely making chains of bubbles.
In the future i will need to modify the pattern detection code to
do stuff like merge chains together.

HOWEVER. for now this actually works and i think the best course of
action is now focusing on getting the whole pipeline of this program
working. from there we can evaluate tweaks to the pattern detection
stuff with minimal hassle.