ENH: reenable toggling selected node/edge/pat info
Also renamed cluster to pattern in this ui. "cluster" was more of
a technical term based on GraphViz history anyway :P

We'll eventually have to figure out a way to populate these tables
with just the stuff in the actual data (so, optional stuff like
coverage, edge multiplicity, GC content, etc.)

but that is a problem for slightly later than now. now is the time
to focus on drawing.