Fix units in Pattern.set_bb(): resolves offset bug
Addresses the first item on #163.

TLDR, widths/heights of patterns were stored in inches while
coordinates were stored in points. YUCK. converting fixes this.

This underscores that I *REALLY* need to spend some time on ensuring
that units throughout the code are consistent. Honestly things now
are a nightmare of some stuff being in points, some stuff being in
inches, and some stuff being in pixels.