ENH: Fix component tiling, kind of
Ideally we'd tile things from top to bottom, but long story short
the way coordinates are stored makes it difficult to do that with
the bounding box stuff. bizarrely, and this seems super
counterintuitive to me so it's probs just me being tired, it's easier
to tile from bottom up due to the nature of how Cytoscape.js
y-coordinates work.

In any case, now components shouldn't overlap. eventually we should
try to use horizontal tiling and stuff but this at least minimally
works now