Infrastructure for selected ele tbls; show patt types
Not sure how I feel about leaving data in the DataHolder and just
doing the iteration over all of that when we add a pattern. It might
get slow -- but then again, storing all of the elements' data in
Cytoscape.js might also be slow. Let's I guess just see how it goes.