Initial commit of collapsing/uncollapsing (broken)
Need to set up patterns for this with the proper data -- basically
replicating initClusters(). incoming and outgoing edge maps, as well
as interior eles (which will live in scratch data since Cy.js eles
are self referential so storing them in data explodes things iirc).

this is ... alarmingly close to being feasible.