temporarily address #231 (#232)
Hmmmm, so is the issue closeable or not?

I guess in the ideal case the data(thickness) edge style doesn't
apply to all edges, just non-dup edges, and then there's no risk
of redundancy or whatevs. but for now this solution works fine
(and i actually kinda line the new way dup edges are styled, they're
like big slinkies -- calls out the fact that they're not real).