DOC: GFA & FASTG parsing and multigraphs #75
Although the parsing functions return nx.MultiDiGraph objects,
they can't actually handle graphs containing duplicate edges.

This is due to the other libraries we are using to parse these
graphs -- GfaPy and pyfastg, respectively.

For FASTG, I'm not sure there is a need to update this behavior,
but for GFA it might be worth addressing this. Although this would
involve making a change to GfaPy, I think -- that or doing some weird
custom stuff to define our own link classes? -- e.g.
https://github.com/ggonnella/gfapy/blob/f150d9e0ca6f82f522bdb5b61669ff27837bd294/gfapy/line/common/connection.py#L99-L110
... So it might be easier to just revert to using an ad hoc line-by-
line GFA parser, rather than GfaPy. But let's not worry about that for
now.