MNT: Rename (decomposed_)digraph to (...)graph
This is the simplest solution I could think of. These graphs are now
MultiDiGraphs, so just calling them "digraphs" could be confusing.
And I think calling them "mdigraphs" or "multidigraphs" or "mdgs"
or whatever would also be confusing, for different reasons.
SOOOO they're graphs, whatever (i think everyone should understand
the inherent generality of that term in this context).

fortunately, it doesn't seem like i have broken any tests with this
change (at least, i haven't broken any tests more than they were
already broken ._.) The next change will be refactoring AssemblyGraph
and the surrounding infrastructure to refer to edges *and their
keys* in the graph (now that this is a multigraph); this will probably
go hand-in-hand with refactoring AssemblyGraph's data model to store
node / edge info in easier-to-access places than the graph objects
(which will help when the hierarchical decomposition stuff starts
breaking things).