MNT/ENH: move logging to AG, and output stats
this way all of the logging done in main.py is based on just the
non-AssemblyGraph stuff. I think consistency makes sense.

It would be nice to make the logs say, for example, how many seconds
have elapsed since execution has started (like in strainFlye).
But that's not needed at all -- just another nice-to-have.

[ci skip]