MNT: frustrating work on decomp stuff
the issue of identifying all chains/bulges in bubbles is annoying,
because it gets kinda tedious to say "after EVERY SINGLE change to
the graph, rerun the bulge / chain checks on the neighbors of the
newly-collapsed pattern." It's a bit inelegant and may break some
tests, but it may be nicer to just extend the superbubble algorithm
minimality check with bulge and chain checking (so that, if you try
to find a superbubble containing a chain or bulge, the validator will
return one of these patterns - not the full bubble).

this is silly, but it GUARANTEES we can't accidentally squish a
chain or bulge into a bubble without first explicitly detecting it.