MNT: refactor Pattern to use Node/Edge obj lists
... rather than referring to an AssemblyGraph object.

This is a huge refactor that touched a lot of code (incl. some only-
tangentially related stuff... sorry historians lol). But this
massively simplifies the Pattern stuff and makes it a lot easier to
test that object, which is super exciting.

Of course, now pygraphviz is segfaulting during layout (?????), so
we have that to move to (lol). But this should make next steps
(final decomposition fixes - chain merging, unneeded split nodes
merging, frayed rope fixes; and layout stuff) much easier

[ci skip]