BUG/MNT: layout fixes
the tests were segfaulting (???) and i thought it was a pygraphviz
issue a la https://github.com/pygraphviz/pygraphviz/issues/113
but it turned out i just forgor to use an f-string so the node
shapes were just being passed as "{node.shape}" or w/e. oof.

also added a space so that the gv_input first line looks like
"digraph thing {" rather than "digraph thing{". not ... not that
that matters, tho, probably.