REL: remove pkg imports from top-level __init__
the reason for this is that we import __version__ from __init__
in setup.py, when we haven't installed mgsc yet. this means that trying
to import something which in turn imports a dependency we haven't
installed yet (in this case, networkx) breaks things.

It doesn't seem like removing these imports in the top-level
__init__.py broke anything, and tbh they were just syntactic sugar
anyway. imo the tradeoff of "having to remember to always update
version number in 2 places" (setup.py, and elsewhere in code when
logging it) vs. "maybe screwing around in the repl is a bit easier
at first" sways towards ... wait what even is this sentence. LOOK
i removed the imports ok. i'm tired