DOC/MNT: rm incorrect comment
we can have a single node be the start and end of a pattern after
removing unnecessary split nodes -- e.g. if the left split and
right split of a pattern are both the start/end of a cyclic chain.
this happens in the chain-in-cyclic-chain-merging test (the bubble
winds up being both the start and end).

SHOOOUld be fine i think, since the start/end stuff is only (iirc)
used for h.decomp stuff anyway

[ci skip]