BUG/MNT: separate (dec) graph traversal in unnecrm
this was causing the august1 graph to fail. now we can find patterns in
that!

the keys were getting out of sync between the graphs, i guess.
this might impact the original splitting code also further up in
the file, but i think it'll just throw an error if we try to remove
an edge that doesn't exist...? so i think that's not urgent, tho
it'd be nice to test things and run over that code with a finer comb

...actually thinking some more, since the dec graph is originally
a copy of the uncollapsed graph, and since during splitting the nodes
being split by definition haven't occurred in any patterns yet, i
thiiiiink that should be safe. but the testing thing really should
happen lol

[ci skip]