ENH/DOC: add DOT CLI option; outputs optional; QOL
lotsa nice changes, e.g. fancy output text and also using "" around
filenames in the logging output (not confident this hits every single
time a filename is logged but we don't have to be perfect)

it's now possible to run mgsc -i [graph] -od [dot] without getting
any errors -- so, the viz stuff is all that's left (well, it's a ton
of work, but still). v nice

aaalso i need to test this new stuff oof