MNT: remove a lotta unused stuff, esp in config.py
yeah, the cool OCD-and-panic-infused function I wrote five years ago
that super duper tries not to accidentally overwrite anything is very
~thoughtful~, but it'll probably just end up breaking on some future
hardware (plus it's ugly to test). I take a more cavalier approach to
this now: if someone is *trying* to get their file overwritten, to the
point where the standard race conditions trigger, then go ahead. Not our
problem. Need to focus effort on the parts of this that matter.