TST: test to_dict() fails if layout not done
this technically fixes the build, since the python tests are now
all passing! (uh, i don't think the js tests are broken at least lol,
not that there are that many of them yet)

note however that this is kind of a cop-out, since i haven't actually
fixed layout() / to_dict() yet regarding the refactor - that's up next
on the agenda (that, or i'll finish implementing the node/edge metadata,
i guess - will be convenient to get that done now while i still remember
how the new data models work).