messing around with graph loading; revert pkg name
went through a whole odyssey today of being like let's just have
the user upload a graph to the flask UI. but then like that would
duplicate the graph on the filesystem, or at least make us
read it slowly unnecessarily, and maybe for like "just" graphs
that's not a big deal but for big files including sequences that
will be a no-no. also the whole point of switching to a desktop-ish
app is supporting big files / sequences so.... no

there are various approaches.

- we can't just get the absolute filepath
  from the browser file select dialog b/c of security issues
  (https://stackoverflow.com/q/76765852 is a kindred spirit),

- we don't want to save the file somewhere (b/c then it is duplicated
  unnecessarily),

- we don't want to iterate through the file once without saving it
  (b/c that will be slow compared to just iterating through the
  "real" version of the file directly rather than having the user
  squeeze it through the browser),

- we COULD make pressing the "load graph" button in the browser
  open a tkinter file selection dialog
  (https://stackoverflow.com/a/14119223) which WOULD give us the
  absolute filepath. this is an extremely funny and also hideous
  workaround. i think this just might work, but the UX (at least
  initially) would be extremely jarring. also if we want to support
  port forwarding eventually, this will not work with that.

- we COULD just go back to what we were doing, and have the user
  specify the graph to MgSc when running the CLI. Should we then
  have other files be passed to the CLI, also? I think sequences
  (e.g. MetaCarvel scaffolds, graph.fa files from LJA, ...) are
  a definite "yes," for the same reasons as with big-graphs-
  containing-sequences above.

  I think the UX here might even be better, tbh, because there's
  no need to navigate a cumbersome file selection dialog >= 2x
  in a row.

  We can think later on about how to take in node/edge metadata,
  and scaffold / AGP files. These will all probs be small enough
  and purely-decorative-to-the-viz enough that we can take them
  in through the browser, but I guess there's not much stopping
  us from taking them in through the CLI initially if users feel
  like doing that.

^^^ so let's do this, I think.

We are left with other questions. I spent some time today thinking
about if using Flask for this is even a good idea... or, more
specifically, if using Flask *directly* for this is a good idea.
The JS code is not really in a "framework," and maybe this is a good
time to learn React? Or Angular or something? In order to make the
code more maintainable.

React + Flask can totally be combined, and this is actually pretty
much what Dash is doing under the hood. Dash is what ModDotPlot
uses, and Dash even has a plugin (?) that integrates it with
Cytoscape.js now. So, should we just switch to using Dash with
Cytoscape.js entirely?

I am not sure. I think it might give us the fastest time-to-some-
sort-of-app, but it seems like it might be cumbersome to adjust
the UI to match what exactly I want. I think sticking to using
Flask directly will make custom stuff easier later on, plus I
think the templating system (and sticking with the shift to
Require.js that I started forever ago) will make the JS code
more maintainable (even if not super cutting-edge, which I am
fine with). Biased opinion from me, I think, but the front-end
code is not really a rat's nest... it's just got a lot of
features in it (or at least it did before I temporarily ripped
them all out during the refactoring process).

So maybe we stick with just Flask for now, and -- as discussed
above -- we have users run MgSc with the graph and seqs as input,
which then just opens up a Flask server in the browser with the
viz stuff on demand.

There are, of course, more questions. Using Flask gives us the
opportunity to set up port forwarding (like ModDotPlot supports),
which would be really cool to support (you could imagine
finding a big asm graph on a server and wanting to viz it natively),
but that comes with its own security implications. There is a
reason that Jupyter Notebooks make you jump through some hoops when
you use them. I think it would be reasonable to set port forwarding
stuff aside for now and worry about supporting it officially later.

These kinds of days are always a bit frustrating. Planning things out
doesn't feel like making progress so much as just... making progress.
But I think this will be a good base for the path forward. eesh.