Draw graph on button click
yay! this works. i think this is sufficient proof that using dash
in this way can work to make the app interactive.

destroying/recreating cy on clicking "draw" is matching what
old mgsc versions used to do. we can argue about whether or not this
is optimal but this demonstrates that it at least works.