move styling from CSS to Python; expand cy
spent SO LONG messing with trying to prevent the canvas from
jumping to the left when we resize the cy div. it is infuriating
how it works so elegantly at
https://js.cytoscape.org/demos/fcose-gene/ but here we have
the janky jumps.

i have a few more ideas but eventually it will be time to either
right-align the control panel like they do (fixes the issue)
or just cut our losses and move on to things that actually matter