treemaps! & fix ctrlSep width & full graph again
so the thing with the treemaps: super cool for smallish graphs
like the E. coli velvet example. but when you get up to like
~100k components, as in the august1 metacarvel graph i have lying
around for testing, this crashes my computer.

so.... we gotta figure out some way of collapsing groups of ccs
together. Or just make these plots not interactive, etc.