pass info to infoDialog stats tab; monospace font
seems good so far. there is jank, i know, in figuring out whether
to count like... self-implying edges of the form +A -> -A towards
these statistics, distinguishing which particular types of graphs
should be considered as "implicit" vs. "explicit," etc.*

but i think this is ok for now - can sweat the deets later as needed

* ok so like... just to write this down to force myself to remember
it... in GFA files, you can have two segments A and B and a
single link from +A -> -B, and this implies the presence of 4
segments (+A/-A and +B/-B) as well as the RC link from +B -> -A.
However, in MetaCarvel GML files, nodes and edges do not (iirc)
have defined RCs, so we just read in the graph as is without adding
on extra stuff.

And there are things in between - like in LJA DOT files all nodes
and edges are explicitly given, but they should still have this
symmetry (like if we see node +A we should also see node -A later
on), even though I don't think the parser code I wrote checks for
this explicitly.

yuck :|

ANYWAYYYYY