save seq lengths in asm graph obj & plot hists
maybe saving these lengths in memory forever will be a bottleneck
in the future, but for now i think it is nbd

Anyway! I think we can safely say that the graph info dialog is now
up and running and useful. Can ofc add more stuff like edges per
component, interactive histograms with dynamic bin sizes a la
https://community.plotly.com/t/jupyter-widget-zoomable-histogram/16849/5,
etc, but for now i think it makes sense to move on to the important
stuff like drawing the graph.