fix cc hist bug (including 0???) & add err chcking
currently this is sloppy - since plot_cc_hist() outputs to
the source of the <img> containing the cc histogram, this writes
the "hey fyi" message as the image source, which doesn't look
like anything.

We COULD still show these histograms for, like, just 1 cc, but I
think that is silly. I think either (1) just showing a <p> with
a clarifying message here if there is 1 cc, or (2) not showing
the Components tab entirely if there is 1 cc, are better options.