add cc node name selector&toggle selection ele viz
this works but tbh the UX of this dropdown-disguising-itself-as-
a-<select> is kind of janky. i might just ditch the icons and go
back to a natural <select> because the old version of mgsc with
a normal <select> just FEELS nicer. like there is a very obvious
lag btwn selecting something and seeing the page change, and
the dropdown doesn't really feel like a selection menu should, ...